make office-map a component
