+++ app/controllers/index.ts
@@ -1,6 +1,5 @@
 import documentText from './documentText';
 import authReturn from './authReturn';
-import contact from './contact';
 import discussionList from './discussionList';
 import hivers from './hivers';
 import meta from './meta';
@@ -10,7 +9,6 @@ import passwordRequest from './passwordRequest';
 export default function(app) {
   documentText(app);
   authReturn(app);
-  contact(app);
   discussionList(app);
   hivers(app);
   meta(app);
