+++ html/contact/contact.html
@@ -72,7 +72,7 @@
     </div>
 
     <!-- Right Row -->
-    <div class="col-md-6 col-xs-11 ph-xl-margin-bottom" ng-controller="ContactMapCtrl">
+    <div class="col-md-6 col-xs-11 ph-xl-margin-bottom">
       <h3 class="ph-heading-light">PaperHive office</h3>
       <address class="sr-only">
         PaperHive,<br/>
@@ -81,7 +81,7 @@
         13355 Berlin,<br />
         Germany
       </address>
-      <leaflet center="center" markers="markers" class="ph-contact-map"></leaflet>
+      <office-map></office-map>
     </div>
 
     <!-- Right Row -->
